+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
-    boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<OrderStatus> status);
+    boolean existsByOrderTableIdInAndOrderStatusIn(List<Long> orderTableIds, List<Order.OrderStatus> status);
 
-    boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<OrderStatus> status);
+    boolean existsByOrderTableIdAndOrderStatusIn(Long orderTableId, List<Order.OrderStatus> status);
 }
