--- src/main/java/kitchenpos/order/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.order.domain;
-
-public enum OrderStatus {
-    COOKING, MEAL, COMPLETION
-}
