+++ src/main/java/kitchenpos/order/service/OrderService.java
@@ -2,9 +2,9 @@ package kitchenpos.order.service;
 
 import java.util.List;
 import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.Order.OrderStatus;
 import kitchenpos.order.domain.OrderLineItemRepository;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
