+++ src/main/java/kitchenpos/table/service/TableGroupValidator.java
@@ -3,8 +3,8 @@ package kitchenpos.table.service;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import kitchenpos.order.domain.Order.OrderStatus;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import org.springframework.stereotype.Component;
