+++ src/main/java/kitchenpos/table/service/TableValidator.java
@@ -1,8 +1,8 @@
 package kitchenpos.table.service;
 
 import java.util.Arrays;
+import kitchenpos.order.domain.Order.OrderStatus;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import org.springframework.stereotype.Component;
