+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -19,9 +19,9 @@ import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.menugroup.service.MenuGroupRequest;
 import kitchenpos.menugroup.service.MenuGroupResponse;
 import kitchenpos.menugroup.service.MenuGroupService;
+import kitchenpos.order.domain.Order.OrderStatus;
 import kitchenpos.order.domain.OrderLineItemRepository;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.service.OrderRequest;
 import kitchenpos.order.service.OrderRequest.OrderLineItemRequest;
 import kitchenpos.order.service.OrderResponse;
