+++ src/test/java/kitchenpos/domain/validator/TableValidatorTest.java
@@ -7,8 +7,8 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
 import java.util.Arrays;
+import kitchenpos.order.domain.Order.OrderStatus;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.service.TableValidator;
 import org.junit.jupiter.api.BeforeEach;
